Title: 특정 숫자 도달하기, Time: 136ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 241016/특정 숫자 도달하기/reaching-specific-number.py
@@ -8,4 +8,4 @@ for i in nums:
         count += 1
     else:
         break
-print(sum(answer), sum(answer)/count)
+print(sum(answer), int(sum(answer)/count))
